Extend max sources number to 65535 (from 255) (close #1859)

+++ application/apps/indexer/session/src/handlers/observing/mod.rs
@@ -38,7 +38,7 @@ pub async fn run_source<S: ByteSource>(
     operation_api: OperationAPI,
     state: SessionStateAPI,
     source: S,
-    source_id: u8,
+    source_id: u16,
     parser: &ParserType,
     rx_sde: Option<SdeReceiver>,
     rx_tail: Option<Receiver<Result<(), tail::Error>>>,
@@ -73,7 +73,7 @@ pub async fn run_source<S: ByteSource>(
 async fn run_producer<T: LogMessage, P: Parser<T>, S: ByteSource>(
     operation_api: OperationAPI,
     state: SessionStateAPI,
-    source_id: u8,
+    source_id: u16,
     mut producer: MessageProducer<T, P, S>,
     mut rx_tail: Option<Receiver<Result<(), tail::Error>>>,
 ) -> OperationResult<()> {
